+++ backend/src/main/java/com/twtw/backend/domain/member/controller/MemberController.java
@@ -1,6 +1,6 @@
-package com.twtw.backend.module.member.controller;
+package com.twtw.backend.domain.member.controller;
 
-import com.twtw.backend.module.member.service.MemberService;
+import com.twtw.backend.domain.member.service.MemberService;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
