+++ backend/src/main/java/com/twtw/backend/domain/member/dto/request/MemberSaveRequest.java
@@ -1,16 +1,13 @@
-package com.twtw.backend.module.member.dto.request;
+package com.twtw.backend.domain.member.dto.request;
 
-import com.twtw.backend.module.member.entity.OAuth2Info;
-import com.twtw.backend.module.member.entity.Role;
+import com.twtw.backend.domain.member.entity.Role;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
 
 @Getter
-@Setter
-@RequiredArgsConstructor
+@AllArgsConstructor
 public class MemberSaveRequest {
     private String nickname;
