+++ backend/src/main/java/com/twtw/backend/domain/member/dto/request/TokenRequest.java
@@ -1,12 +1,12 @@
-package com.twtw.backend.module.member.dto.request;
+package com.twtw.backend.domain.member.dto.request;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 @Getter
-@Setter
-@RequiredArgsConstructor
+@AllArgsConstructor
 public class TokenRequest {
     String accessToken;
     String refreshToken;
