+++ backend/src/main/java/com/twtw/backend/domain/member/dto/response/TokenDto.java
@@ -1,11 +1,10 @@
-package com.twtw.backend.module.member.dto.response;
+package com.twtw.backend.domain.member.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
-@Setter
 @AllArgsConstructor
 public class TokenDto {
     private String accessToken;
