+++ backend/src/main/java/com/twtw/backend/domain/member/entity/AuthType.java
@@ -1,0 +1,6 @@
+package com.twtw.backend.domain.member.entity;
+
+public enum AuthType {
+    APPLE,
+    KAKAO
+}
