+++ backend/src/main/java/com/twtw/backend/domain/member/entity/Role.java
@@ -1,0 +1,6 @@
+package com.twtw.backend.domain.member.entity;
+
+public enum Role {
+    ROLE_ADMIN,
+    ROLE_USER
+}
