--- backend/src/main/java/com/twtw/backend/module/member/entity/AuthType.java
@@ -1,9 +1,0 @@
-package com.twtw.backend.module.member.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum AuthType {
-    APPLE,
-    KAKAO
-}
