--- backend/src/main/java/com/twtw/backend/module/member/entity/Role.java
@@ -1,9 +1,0 @@
-package com.twtw.backend.module.member.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum Role {
-    ROLE_ADMIN,
-    ROLE_USER
-}
