+++ backend/src/main/resources/application.yml
@@ -1,5 +1,3 @@
-server:
-  port: 8080
 spring:
   profiles:
     include: oauth
